Add an animation redraw cycle
This retains the subtree but will recompute attributes on
config.modifier.  This should let us fast track animations without
needing to rerun the potentially expensive user defined view functions.

This could still be slow if the user writes an expensive modifier.  The
user should always be careful to profile modifiers since they will be
the be among the most run functions in a module.